# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Family Budget Service API
  version: 1.0.0
  description: |
    Целевой контракт `/api/v1` для Android-клиента (spec 005, решения A-01…A-09).

    Код и спецификация совпадают: `tests/integration/openapi_coverage_test.go` проверяет
    обе стороны — и зарегистрированный роут без описания здесь, и описанную операцию без
    роута роняют `make test`.

    Соглашения:
    - Деньги — целые в минимальных единицах валюты семьи (`amount_minor`, int64). Валюта — у семьи (A-05).
    - Дата операции — календарная `YYYY-MM-DD` без времени; служебные метки — RFC3339 UTC (A-06).
    - `POST` для транзакций, бюджетов и категорий принимает клиентский `id` (любой валидный UUID);
      повтор с тем же `id` возвращает `200` с существующей записью, тело при этом игнорируется (A-07).
    - Дробные числа в двух единицах: доли 0…1 — `share`, `*_delta` и `utilization` в
      `StatsSummary.budgets[]` (`BudgetProgress`); проценты 0…100 — `percentage` и `utilization`
      в `Budget` (`GET /api/v1/budgets`).
    - Списки: `limit` (по умолчанию 50, максимум 200) и `offset`; `meta.pagination.total` — общее число (A-08).
    - Ошибка всегда одной формы: `{"error": {...}, "meta": {...}}` (A-08).

    Коды ошибок: `INVALID_REQUEST` / `INVALID_ID` (400 — тело или id не разобрались),
    `REQUEST_TIMEOUT` (408 — тело не успело загрузиться), `PAYLOAD_TOO_LARGE` (413),
    `UNAUTHORIZED` (401), `FORBIDDEN` (403), `NOT_FOUND` / `<ENTITY>_NOT_FOUND` (404),
...
  license:
    name: MIT
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: Счета семьи
      description: admin и member. Порядок — по имени без учёта регистра.
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - name: archived
          in: query
          required: false
          description: '`true` добавляет в список архивные счета'
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Список счетов
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Account' }
                  meta: { $ref: '#/components/schemas/ListMeta' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    post:
      operationId: createAccount
      tags: [accounts]
      summary: Создать счёт
      description: |
        admin и member. Повтор с тем же `id` возвращает `200` с существующим счётом; остальное тело
        при этом игнорируется (A-07). Имя уникально в семье без учёта регистра и пробелов по краям,
        архивный счёт его тоже занимает — `409 ACCOUNT_NAME_EXISTS`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAccountRequest' }
      responses:
        '201': { $ref: '#/components/responses/AccountOk' }
        '200': { $ref: '#/components/responses/AccountOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/accounts/{id}:
    parameters:
      - { $ref: '#/components/parameters/PathID' }
    put:
      operationId: updateAccount
      tags: [accounts]
      summary: Переименовать, архивировать или вернуть счёт
      description: admin и member. Архивный счёт скрыт из списка без `?archived=true` и не назначается
        новым операциям.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateAccountRequest' }
      responses:
        '200': { $ref: '#/components/responses/AccountOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Удалить счёт
      description: |
        Только admin. Счёт с операциями или сверками не удаляется — `409 ACCOUNT_IN_USE`;
        перевыпущенную карту архивируют.
      responses:
        '204': { description: Счёт удалён }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/holdings:
    get:
      operationId: listHoldings
      tags: [holdings]
      summary: Активы и пассивы семьи
      description: |
        admin и member. Порядок — по имени без учёта регистра. `current` — последний снимок с датой не позже
        сегодняшнего дня в часовом поясе семьи, `null` — снимков нет.
      parameters:
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
        - name: archived
          in: query
          required: false
          description: '`true` добавляет в список архивные позиции'
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Список позиций
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Holding' }
                  meta: { $ref: '#/components/schemas/ListMeta' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    post:
      operationId: createHolding
      tags: [holdings]
      summary: Создать актив или пассив
      description: |
        admin и member. Повтор с тем же `id` возвращает `200` с существующей позицией; остальное тело
        при этом игнорируется (A-07). Имя уникально в семье без учёта регистра и пробелов по краям,
        архивная позиция его тоже занимает — `409 HOLDING_NAME_EXISTS`. `kind` не из списка своей
        стороны — `422`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateHoldingRequest' }
      responses:
        '201': { $ref: '#/components/responses/HoldingOk' }
        '200': { $ref: '#/components/responses/HoldingOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
  /api/v1/holdings/{id}:
    parameters:
      - { $ref: '#/components/parameters/PathID' }
    put:
      operationId: updateHolding
      tags: [holdings]
      summary: Переименовать, сменить вид, архивировать или вернуть позицию
      description: |
        admin и member. `side` не меняется: поля в запросе нет, присланное игнорируется. Архив только
        прячет позицию из списка без `?archived=true` — в капитале она остаётся по своим снимкам;
        прекращение стоимости — снимок `0`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateHoldingRequest' }
      responses:
        '200': { $ref: '#/components/responses/HoldingOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      operationId: deleteHolding
      tags: [holdings]
      summary: Удалить позицию
      description: |
        Только admin. Снимки удаляются вместе с позицией, и прошлые значения капитала меняются — это
        исправление ошибочно заведённой позиции. Проданное имущество — снимок `0` и архив.
      responses:
        '204': { description: Позиция удалена }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Непрозрачный токен из `POST /api/v1/auth/login` (A-01): 32 байта `crypto/rand` в base64url.
        Не JWT — сервер хранит SHA-256 и проверяет срок и активность пользователя на каждом запросе.
        Скользящий срок — 30 дней без активности, потолок — 180 дней от выдачи.
  parameters:
    PathID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
  responses:
    BadRequest:
      description: |
        Тело или идентификатор не разобрались: `INVALID_REQUEST` (сломанный JSON или неверный
        тип поля), `INVALID_ID` (в пути не UUID). Ошибки валидации значений — это `422`.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Токена нет, он истёк или отозван (`UNAUTHORIZED`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Forbidden:
      description: Роль не даёт доступа к операции (`FORBIDDEN`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: |
        Объект не найден: `NOT_FOUND` для неизвестного пути, `<ENTITY>_NOT_FOUND`
        (`USER_NOT_FOUND`, `SESSION_NOT_FOUND`, `CATEGORY_NOT_FOUND`, …) для отсутствующей записи
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: |
        Состояние не допускает операцию: `SETUP_REQUIRED`, `CURRENCY_LOCKED`, `LAST_ADMIN`,
        `CANNOT_DEACTIVATE_SELF`, `EMAIL_TAKEN`, `BUDGET_OVERLAP`, `BUDGET_NAME_EXISTS`,
        `BUDGET_BELOW_SPENT`, `BUDGET_ID_EXISTS`, `BUDGET_NOT_TAIL`, `ACCOUNT_NAME_EXISTS`,
        `ACCOUNT_IN_USE`, `HOLDING_NAME_EXISTS`
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationFailed:
      description: Тело или параметры не прошли валидацию (`VALIDATION_ERROR`); поля — в `error.details`
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    AccountOk:
      description: Счёт
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AccountOk' }
    HoldingOk:
      description: Позиция
      content:
        application/json:
          schema: { $ref: '#/components/schemas/HoldingOk' }
...
  schemas:
    Money:
      type: integer
      format: int64
      maximum: 99999999999
      description: |
        Сумма в минимальных единицах валюты семьи (копейки для RUB, центы для EUR/USD).
        Поддерживаются только валюты с двумя знаками после запятой (A-05).
      examples: [125000]
    CalendarDate:
      type: string
      format: date
      description: Календарная дата без времени в часовом поясе семьи (A-06)
      examples: ['2026-09-04']
    Timestamp:
      type: string
      format: date-time
      description: Служебная метка времени, RFC3339 UTC
    Meta:
      type: object
      required: [request_id, timestamp]
      properties:
        request_id:
          type: string
          description: Значение заголовка `X-Request-ID` — тот же id в логах сервера
        timestamp: { $ref: '#/components/schemas/Timestamp' }
        version:
          type: string
          description: Версия сборки, как в `/health`
        pagination:
          $ref: '#/components/schemas/Pagination'
    ListMeta:
      allOf:
        - { $ref: '#/components/schemas/Meta' }
        - type: object
          required: [pagination]
          properties:
            pagination: { $ref: '#/components/schemas/Pagination' }
    Pagination:
      type: object
      required: [limit, offset, total]
      description: Присутствует только в ответах-списках (A-08)
      properties:
        limit: { type: integer, minimum: 1, maximum: 200 }
        offset: { type: integer, minimum: 0 }
        total:
          type: integer
          description: Общее число записей, подходящих под фильтр, без учёта limit/offset
    Error:
      type: object
      required: [error, meta]
      description: Единственная форма ошибки во всём API (A-08)
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Машиночитаемый код, `SCREAMING_SNAKE_CASE`
              examples: ['VALIDATION_ERROR']
            message:
              type: string
              description: Сообщение для человека; не предназначено для показа пользователю без перевода
            details:
              type: array
              description: Заполняется при `VALIDATION_ERROR` — по элементу на поле
              items: { $ref: '#/components/schemas/ErrorDetail' }
        meta: { $ref: '#/components/schemas/Meta' }
    ErrorDetail:
      type: object
      required: [field, message, code]
      properties:
        field:
          type: string
          description: Путь к полю в теле запроса или имя query-параметра
          examples: ['amount_minor']
        message: { type: string }
        code:
          type: string
          examples: ['required']
...
    AccountOk:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/Account' }
        meta: { $ref: '#/components/schemas/Meta' }
    HoldingOk:
      type: object
      required: [data, meta]
      properties:
        data: { $ref: '#/components/schemas/Holding' }
        meta: { $ref: '#/components/schemas/Meta' }
...
    Account:
      type: object
      required: [id, name, is_archived, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        is_archived: { type: boolean }
        created_at: { $ref: '#/components/schemas/Timestamp' }
        updated_at: { $ref: '#/components/schemas/Timestamp' }
    CreateAccountRequest:
      type: object
      required: [name]
      properties:
        id:
          type: string
          format: uuid
          description: Необязательный клиентский UUID для идемпотентного повтора; тело повтора игнорируется (A-07)
        name:
          type: string
          maxLength: 50
          description: Пробелы по краям обрезаются; пустое после обрезки — `422`
    UpdateAccountRequest:
      type: object
      minProperties: 1
      description: Хотя бы одно поле обязательно
      properties:
        name: { type: string, maxLength: 50 }
        is_archived: { type: boolean }
    HoldingSide:
      type: string
      enum: [asset, liability]
      description: Знак вклада в капитал; задаётся при создании и не меняется
    HoldingKind:
      type: string
      description: |
        Вид позиции. Активы — `cash`, `deposit`, `investment`, `property`, `vehicle`, `other`;
        пассивы — `mortgage`, `loan`, `credit_card`, `other`. Не `enum`: список может расти, и
        неизвестный вид клиент показывает как `other`.
    Holding:
      type: object
      required: [id, name, side, kind, is_archived, current, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        side: { $ref: '#/components/schemas/HoldingSide' }
        kind: { $ref: '#/components/schemas/HoldingKind' }
        is_archived: { type: boolean }
        current:
          oneOf:
            - { $ref: '#/components/schemas/HoldingCurrent' }
            - { type: 'null' }
          description: Последний снимок не позже сегодняшнего дня семьи; `null` — снимков нет
        created_at: { $ref: '#/components/schemas/Timestamp' }
        updated_at: { $ref: '#/components/schemas/Timestamp' }
    HoldingCurrent:
      type: object
      required: [date, value_minor]
      properties:
        date: { $ref: '#/components/schemas/CalendarDate' }
        value_minor:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          minimum: 0
    CreateHoldingRequest:
      type: object
      required: [name, side, kind]
      properties:
        id:
          type: string
          format: uuid
          description: Необязательный клиентский UUID для идемпотентного повтора; тело повтора игнорируется (A-07)
        name:
          type: string
          maxLength: 50
          description: Пробелы по краям обрезаются; пустое после обрезки — `422`
        side: { $ref: '#/components/schemas/HoldingSide' }
        kind: { $ref: '#/components/schemas/HoldingKind' }
    UpdateHoldingRequest:
      type: object
      minProperties: 1
      description: Хотя бы одно поле обязательно; `side` не меняется
      properties:
        name: { type: string, maxLength: 50 }
        kind: { $ref: '#/components/schemas/HoldingKind' }
        is_archived: { type: boolean }
...